+++ CHANGELOG.md
@@ -18,6 +18,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 * implement `Default` for `Element` and `Attribute`
 * `attr::disabled()` and `attr::autofocus()`
 * `attr::charset_utf8()` and `elt::meta_charset_utf8()`
+* `attr::autocomplete(type_: impl Into<Cow<'static, str>>)`, `attr::autocomplete_on()`, `attr::autocomplete_off()`
 
 
 ## [1.4.0] - 2024-12-13